refactor(): move code arround into smaller pieces
