+++ src/test/java/com/axway/gw/es/yaml/it/BatchYamlStoreConverterTestIT.java
@@ -6,15 +6,18 @@ import java.util.Map;
 
 import com.axway.gw.es.yaml.YamlEntityStore;
 import com.axway.gw.es.yaml.YamlPK;
-import com.axway.gw.es.yaml.converters.ConvertToYamlStore;
-import com.axway.gw.es.yaml.converters.EntityDTOConverter;
+import com.axway.gw.es.yaml.converters.EntityStoreESPKMapper;
+import com.axway.gw.es.yaml.tools.ConvertToYamlStore;
+import com.axway.gw.es.yaml.util.NameUtils;
 import com.axway.gw.es.yaml.utils.ESDiff;
+import com.vordel.es.ESPK;
 import com.vordel.es.EntityStore;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 import static com.axway.gw.es.yaml.utils.ESTestsUtil.assertDiffCount;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.InstanceOfAssertFactories.atomicMarkableReference;
 
 public class BatchYamlStoreConverterTestIT {
 
@@ -29,21 +32,34 @@ public class BatchYamlStoreConverterTestIT {
             "TeamDevelopmentSettingsAPIM"})
     public void convertFactoryTemplate(String project) throws InterruptedException, IOException {
 
-        final ConvertToYamlStore convertToYamlStore = new ConvertToYamlStore("federated:file:" + BatchYamlStoreConverterTestIT.class.getResource("/apiprojects/" + project + "/configs.xml").getFile());
-        convertToYamlStore.convert(YAML_OUTPUT_DIR + project);
-       //  convertToYamlStore.writeFederatedToYamlPkMapping(YAML_OUTPUT_DIR + project);
+        final String yamlDir = YAML_OUTPUT_DIR + project;
 
-        final EntityStore sourceES = convertToYamlStore.getInputEntityStore();
+        final ConvertToYamlStore convertToYamlStore = new ConvertToYamlStore(getFEDFileURL(project));
 
+        // convert => All YML Files are written as as the key mapping file
+        convertToYamlStore.convert(yamlDir, true);
+
+        // load all YAML filed
         final YamlEntityStore yamlEntityStore = new YamlEntityStore();
-        yamlEntityStore.connect(YamlEntityStore.SCHEME + "file:" + Paths.get(YAML_OUTPUT_DIR + project).toFile().getPath(), null);
+        yamlEntityStore.connect(YamlEntityStore.SCHEME + "file:" + Paths.get(yamlDir).toFile().getPath(), null);
 
-        // final Map<String, YamlPK> federatedToYamlPk = EntityDTOConverter.readFederatedToYamlPkMapping(YAML_OUTPUT_DIR + project);
+        // Read the key mapping
+        final Map<String, String> keyMapping = new EntityStoreESPKMapper<String, String>().readFederatedToYamlPkMapping(yamlDir);
 
-        ESDiff diff = ESDiff.diff(sourceES, yamlEntityStore, YamlPK::new);
+        // compare using the key mapping to find matching elements
+        ESDiff diff = ESDiff.diff(convertToYamlStore.getInputEntityStore(), yamlEntityStore, fedKey -> {
+            final String path = keyMapping.get(fedKey);
+            if (path == null) return null;
+            final ESPK rootPK = yamlEntityStore.getRootPK();
+            return new YamlPK(rootPK, path);
+        });
 
-        assertDiffCount(diff,0, project);
+        assertDiffCount(diff, 0, project);
 
+    }
+
+    private String getFEDFileURL(String project) {
+        return "federated:file:" + BatchYamlStoreConverterTestIT.class.getResource("/apiprojects/" + project + "/configs.xml").getFile();
     }
